Add scripts to produce density/temp at latest time
This script is still a work in progress.  Right now
I've gotten it to calculate the corresponding
HyDomainAtTime*.txt file for the latest
VolumeMatterData/*.h5 files available.

I got pretty far into it until I realized bash
couldn't do arithmetic of two decimal numbers, so we
wrote a python script to do that in its place.  I'd
probably just want to do this all in python if I
started again..