Add method to autoextract paraview data singletons
With per-segment-volume-generator.sh, you can run
this script on a specifed (relative) path to
extract paraview data on a single segment.

Right now, this reuses some code from the volume-
gen script, which assumes you just want to pull
from the latest time available in
VolumeMatterData/.  Would be good to add an input
option to allow user to specify the time from a
list of times available.